requirements.txt
Flask==3.1.1
Flask-WTF==1.2.2
WTForms==3.0.1
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.38
python-dotenv
pytz